requirements.txt
numpy
pycollada
nilearn
coremltools
torch
torchvision
scipy